get username by email
retrieve the username of a teammate using their email address.

While inviting a user through the SendGrid API, an email address is necessary. However, to retrieve user information, the username becomes essential.

The intention is to ensure uniqueness by obtaining the username from the email address when utilizing this library with terraform-provider.